backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
websockets>=12.0
httpx>=0.25.0
